google-adk/requirements.txt
litellm
google-adk
dotenv
google.genai
uv
